% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array2df.R
\name{array2df}
\alias{array2df}
\title{Array to long-format data frame}
\usage{
array2df(arr, value_name = "value")
}
\arguments{
\item{arr}{An array of N-dimensiones}

\item{value_name}{Name for the column containing the values of the array.}
}
\description{
Organizes the information in an array with N dimensions into a
long-format data frame with N + 1 columns, where the last column
contains the values of the array.
}
\author{
Jose M. Pavia, \email{pavia@uv.es}
}
